requirements.txt
flask>=3.0.0
python-pptx>=0.6.21
openai>=1.3.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
python-dotenv>=1.0.0
